put table in component and use for repairItems
